+++ simple-stack/src/main/java/com/zhuinden/simplestack/navigator/BackstackHost.java
@@ -22,6 +22,7 @@ import android.view.ViewGroup;
 
 import com.zhuinden.simplestack.Backstack;
 import com.zhuinden.simplestack.BackstackManager;
+import com.zhuinden.simplestack.KeyFilter;
 import com.zhuinden.simplestack.KeyParceler;
 import com.zhuinden.simplestack.StateChanger;
 import com.zhuinden.statebundle.StateBundle;
@@ -42,6 +43,7 @@ public final class BackstackHost
 
     StateChanger stateChanger;
 
+    KeyFilter keyFilter;
     KeyParceler keyParceler;
     BackstackManager.StateClearStrategy stateClearStrategy;
     boolean shouldPersistContainerChild;
@@ -62,6 +64,7 @@ public final class BackstackHost
     Backstack initialize(boolean isInitializeDeferred) {
         if(backstackManager == null) {
             backstackManager = new BackstackManager();
+            backstackManager.setKeyFilter(keyFilter);
             backstackManager.setKeyParceler(keyParceler);
             backstackManager.setStateClearStrategy(stateClearStrategy);
             backstackManager.setup(initialKeys);