+++ simple-stack/src/main/java/com/zhuinden/simplestack/navigator/DefaultStateChanger.java
@@ -43,11 +43,65 @@ public final class DefaultStateChanger
         }
     }
 
-    private static class NoOpViewChangeCompletionListener implements ViewChangeCompletionListener {
+    private static class NoOpViewChangeStartListener
+            implements ViewChangeStartListener {
+        @Override
+        public void handleViewChangeStart(@NonNull StateChange stateChange, @NonNull ViewGroup container, @Nullable View previousView, @NonNull View newView, @NonNull Callback startCallback) {
+            startCallback.startViewChange();
+        }
+    }
+
+    private static class NoOpViewChangeCompletionListener
+            implements ViewChangeCompletionListener {
         @Override
         public void handleViewChangeComplete(@NonNull StateChange stateChange, @NonNull ViewGroup container, @Nullable View previousView, @NonNull View newView, @NonNull Callback completionCallback) {
             completionCallback.viewChangeComplete();
         }
+    }
+
+    private static class DefaultLayoutInflationStrategy
+            implements LayoutInflationStrategy {
+        @Override
+        public void inflateLayout(StateChange stateChange, StateKey key, Context context, ViewGroup container, Callback callback) {
+            final View newView = LayoutInflater.from(context).inflate(key.layout(), container, false);
+            callback.layoutInflationComplete(newView);
+        }
+    }
+
+    private static class NavigatorStatePersistenceStrategy
+            implements StatePersistenceStrategy {
+        @Override
+        public void persistViewToState(@NonNull Object previousKey, @NonNull View previousView) {
+            Navigator.persistViewToState(previousView);
+        }
+
+        @Override
+        public void restoreViewFromState(@NonNull Object newKey, @NonNull View newView) {
+            Navigator.restoreViewFromState(newView);
+        }
+    }
+
+    /**
+     * Allows the possibility of listening to when the view change is to start, after the views are inflated and state is restored, but before view change is started.
+     */
+    public interface ViewChangeStartListener {
+        /**
+         * Notifies the {@link DefaultStateChanger} that the view change start listener completed its callback.
+         */
+        public interface Callback {
+            void startViewChange();
+        }
+
+        /**
+         * Called when a view change is to be started.
+         *
+         * @param stateChange   the state change
+         * @param container     the container
+         * @param previousView  the previous view
+         * @param newView       the new view
+         * @param startCallback the start callback that must be called once the view change callback is completed by the listener
+         */
+        void handleViewChangeStart(@NonNull StateChange stateChange, @NonNull ViewGroup container, @Nullable View previousView, @NonNull View newView, @NonNull ViewChangeStartListener.Callback startCallback);
     }
 
     /**
@@ -64,13 +118,36 @@ public final class DefaultStateChanger
         /**
          * Called when a view change is completed.
          *
-         * @param stateChange the state change
-         * @param container the container
-         * @param previousView the previous view
-         * @param newView the new view
+         * @param stateChange        the state change
+         * @param container          the container
+         * @param previousView       the previous view
+         * @param newView            the new view
          * @param completionCallback the completion callback that must be called once the view change callback is completed by the listener
          */
         void handleViewChangeComplete(@NonNull StateChange stateChange, @NonNull ViewGroup container, @Nullable View previousView, @NonNull View newView, @NonNull Callback completionCallback);
+    }
+
+    /**
+     * Allows the possibility of using a custom layout inflation strategy for inflating the new view.
+     */
+    public interface LayoutInflationStrategy {
+        /**
+         * This callback must be called to provide the inflated view.
+         */
+        public interface Callback {
+            void layoutInflationComplete(View view);
+        }
+
+        /**
+         * This method needs to inflate the new view, preferably using the provided context.
+         *
+         * @param stateChange the state change
+         * @param key         the new key this view is inflated for
+         * @param context     the context the layout inflater is originally acquired from
+         * @param container   the container
+         * @param callback    the inflation callback that must be called when layout inflation is complete
+         */
+        void inflateLayout(StateChange stateChange, StateKey key, Context context, ViewGroup container, Callback callback);
     }
 
     /**
@@ -80,7 +157,7 @@ public final class DefaultStateChanger
         /**
          * Persists the previous active view's view state.
          *
-         * @param previousKey the previous key
+         * @param previousKey  the previous key
          * @param previousView the previous view
          */
         void persistViewToState(@NonNull Object previousKey, @NonNull View previousView);
@@ -88,22 +165,10 @@ public final class DefaultStateChanger
         /**
          * Restores the new active view's view state.
          *
-         * @param newKey the new key
+         * @param newKey  the new key
          * @param newView the new view
          */
         void restoreViewFromState(@NonNull Object newKey, @NonNull View newView);
-    }
-
-    private static class NavigatorStatePersistenceStrategy implements StatePersistenceStrategy {
-        @Override
-        public void persistViewToState(@NonNull Object previousKey, @NonNull View previousView) {
-            Navigator.persistViewToState(previousView);
-        }
-
-        @Override
-        public void restoreViewFromState(@NonNull Object newKey, @NonNull View newView) {
-            Navigator.restoreViewFromState(newView);
-        }
     }
 
     private static final NoOpViewChangeHandler NO_OP_VIEW_CHANGE_HANDLER = new NoOpViewChangeHandler();
@@ -111,7 +176,9 @@ public final class DefaultStateChanger
     private Context baseContext;
     private ViewGroup container;
     private StateChanger externalStateChanger;
+    private ViewChangeStartListener viewChangeStartListener;
     private ViewChangeCompletionListener viewChangeCompletionListener;
+    private LayoutInflationStrategy layoutInflationStrategy;
     private StatePersistenceStrategy statePersistenceStrategy;
 
     /**
@@ -122,7 +189,9 @@ public final class DefaultStateChanger
      */
     public static class Configurer {
         StateChanger externalStateChanger = null;
+        ViewChangeStartListener viewChangeStartListener = null;
         ViewChangeCompletionListener viewChangeCompletionListener = null;
+        LayoutInflationStrategy layoutInflationStrategy = null;
         StatePersistenceStrategy statePersistenceStrategy = null;
 
         private Configurer() {
@@ -139,6 +208,20 @@ public final class DefaultStateChanger
                 throw new NullPointerException("If set, external state changer cannot be null!");
             }
             this.externalStateChanger = stateChanger;
+            return this;
+        }
+
+        /**
+         * Sets the {@link ViewChangeStartListener}. It is executed before the view change.
+         *
+         * @param viewChangeStartListener the view change start listener
+         * @return the configurer
+         */
+        public Configurer setViewChangeStartListener(@NonNull ViewChangeStartListener viewChangeStartListener) {
+            if(viewChangeStartListener == null) {
+                throw new NullPointerException("If set, view change start listener cannot be null!");
+            }
+            this.viewChangeStartListener = viewChangeStartListener;
             return this;
         }
 
@@ -171,6 +254,20 @@ public final class DefaultStateChanger
         }
 
         /**
+         * Sets the {@link LayoutInflationStrategy}. It is used to inflate the new view before a view change.
+         *
+         * @param layoutInflationStrategy the layout inflation strategy
+         * @return the configurer
+         */
+        public Configurer setLayoutInflationStrategy(@NonNull LayoutInflationStrategy layoutInflationStrategy) {
+            if(layoutInflationStrategy == null) {
+                throw new NullPointerException("If set, layout inflation strategy cannot be null!");
+            }
+            this.layoutInflationStrategy = layoutInflationStrategy;
+            return this;
+        }
+
+        /**
          * Creates the {@link DefaultStateChanger} with the specified parameters.
          *
          * @param baseContext the base context used to inflate the views
@@ -178,7 +275,12 @@ public final class DefaultStateChanger
          * @return the new {@link DefaultStateChanger}
          */
         public DefaultStateChanger create(Context baseContext, ViewGroup container) {
-            return new DefaultStateChanger(baseContext, container, externalStateChanger, viewChangeCompletionListener, statePersistenceStrategy);
+            return new DefaultStateChanger(baseContext,
+                    container,
+                    externalStateChanger,
+                    viewChangeStartListener,
+                    viewChangeCompletionListener, layoutInflationStrategy,
+                    statePersistenceStrategy);
         }
     }
 
@@ -202,10 +304,10 @@ public final class DefaultStateChanger
      * @return the state changer
      */
     public static DefaultStateChanger create(Context baseContext, ViewGroup container) {
-        return new DefaultStateChanger(baseContext, container, null, null, null);
+        return new DefaultStateChanger(baseContext, container, null, null, null, null, null);
     }
 
-    DefaultStateChanger(@NonNull Context baseContext, @NonNull ViewGroup container, @Nullable StateChanger externalStateChanger, @Nullable ViewChangeCompletionListener viewChangeCompletionListener, @Nullable StatePersistenceStrategy statePersistenceStrategy) {
+    DefaultStateChanger(@NonNull Context baseContext, @NonNull ViewGroup container, @Nullable StateChanger externalStateChanger, ViewChangeStartListener viewChangeStartListener, @Nullable ViewChangeCompletionListener viewChangeCompletionListener, LayoutInflationStrategy layoutInflationStrategy, @Nullable StatePersistenceStrategy statePersistenceStrategy) {
         if(baseContext == null) {
             throw new NullPointerException("baseContext cannot be null");
         }
@@ -218,10 +320,18 @@ public final class DefaultStateChanger
             externalStateChanger = new NoOpStateChanger();
         }
         this.externalStateChanger = externalStateChanger;
+        if(viewChangeStartListener == null) {
+            viewChangeStartListener = new NoOpViewChangeStartListener();
+        }
+        this.viewChangeStartListener = viewChangeStartListener;
         if(viewChangeCompletionListener == null) {
             viewChangeCompletionListener = new NoOpViewChangeCompletionListener();
         }
         this.viewChangeCompletionListener = viewChangeCompletionListener;
+        if(layoutInflationStrategy == null) {
+            layoutInflationStrategy = new DefaultLayoutInflationStrategy();
+        }
+        this.layoutInflationStrategy = layoutInflationStrategy;
         if(statePersistenceStrategy == null) {
             statePersistenceStrategy = new NavigatorStatePersistenceStrategy();
         }
@@ -229,12 +339,16 @@ public final class DefaultStateChanger
     }
 
     private void finishStateChange(StateChange stateChange, ViewGroup container, View previousView, View newView, final Callback completionCallback) {
-        viewChangeCompletionListener.handleViewChangeComplete(stateChange, container, previousView, newView, new ViewChangeCompletionListener.Callback() {
-            @Override
-            public void viewChangeComplete() {
-                completionCallback.stateChangeComplete();
-            }
-        });
+        viewChangeCompletionListener.handleViewChangeComplete(stateChange,
+                container,
+                previousView,
+                newView,
+                new ViewChangeCompletionListener.Callback() {
+                    @Override
+                    public void viewChangeComplete() {
+                        completionCallback.stateChangeComplete();
+                    }
+                });
     }
 
     @Override
@@ -278,37 +392,49 @@ public final class DefaultStateChanger
      * @param <T>                the type of the previous key
      * @param <U>                the type of the new key
      */
-    public final <T extends StateKey, U extends StateKey> void performViewChange(T previousKey, U newKey, final StateChange stateChange, final int direction, final Callback completionCallback) {
+    public final <T extends StateKey, U extends StateKey> void performViewChange(final T previousKey, final U newKey, final StateChange stateChange, final int direction, final Callback completionCallback) {
         final View previousView = container.getChildAt(0);
         if(previousView != null && previousKey != null) {
             statePersistenceStrategy.persistViewToState(previousKey, previousView);
         }
         Context newContext = stateChange.createContext(baseContext, newKey);
-        final View newView = LayoutInflater.from(newContext).inflate(newKey.layout(), container, false);
-        statePersistenceStrategy.restoreViewFromState(newKey, newView);
-
-        if(previousView == null) {
-            container.addView(newView);
-            finishStateChange(stateChange, container, previousView, newView, completionCallback);
-        } else {
-            final ViewChangeHandler viewChangeHandler;
-            if(direction == StateChange.FORWARD) {
-                viewChangeHandler = newKey.viewChangeHandler();
-            } else if(previousKey != null && direction == StateChange.BACKWARD) {
-                viewChangeHandler = previousKey.viewChangeHandler();
-            } else {
-                viewChangeHandler = NO_OP_VIEW_CHANGE_HANDLER;
+        layoutInflationStrategy.inflateLayout(stateChange, newKey, newContext, container, new LayoutInflationStrategy.Callback() {
+            @Override
+            public void layoutInflationComplete(final View newView) {
+                statePersistenceStrategy.restoreViewFromState(newKey, newView);
+                viewChangeStartListener.handleViewChangeStart(stateChange,
+                        container,
+                        previousView,
+                        newView,
+                        new ViewChangeStartListener.Callback() {
+                            @Override
+                            public void startViewChange() {
+                                if(previousView == null) {
+                                    container.addView(newView);
+                                    finishStateChange(stateChange, container, previousView, newView, completionCallback);
+                                } else {
+                                    final ViewChangeHandler viewChangeHandler;
+                                    if(direction == StateChange.FORWARD) {
+                                        viewChangeHandler = newKey.viewChangeHandler();
+                                    } else if(previousKey != null && direction == StateChange.BACKWARD) {
+                                        viewChangeHandler = previousKey.viewChangeHandler();
+                                    } else {
+                                        viewChangeHandler = NO_OP_VIEW_CHANGE_HANDLER;
+                                    }
+                                    viewChangeHandler.performViewChange(container,
+                                            previousView,
+                                            newView,
+                                            direction,
+                                            new ViewChangeHandler.CompletionCallback() {
+                                                @Override
+                                                public void onCompleted() {
+                                                    finishStateChange(stateChange, container, previousView, newView, completionCallback);
+                                                }
+                                            });
+                                }
+                            }
+                        });
             }
-            viewChangeHandler.performViewChange(container,
-                    previousView,
-                    newView,
-                    direction,
-                    new ViewChangeHandler.CompletionCallback() {
-                        @Override
-                        public void onCompleted() {
-                            finishStateChange(stateChange, container, previousView, newView, completionCallback);
-                        }
-                    });
-        }
+        });
     }
 }